[fronius] Move battery settings channels and battery control actions to the battery thing
As suggested in the review: the battery thing is the natural home for the
battery settings channels and the battery control actions, now that it
exists. The powerinverter thing keeps the powerflow-based channels
(powerflowbackupmode, powerflowbatterystandby), since they come from the
power flow API it polls.

- The channel ids follow the naming of the existing battery thing channels
  (camelCase, no battery prefix): socMin, socMax, backupReservedCapacity,
  backupCriticalSoc, chargeFromGrid, calibration, nightPreservationLimit.
  This also resolves the chargefromgrid naming review comment.
- The battery thing gets thingTypeVersion 1 with update instructions adding
  the new channels; the powerinverter v2 instructions only add the powerflow
  channels. Since none of these channels have been released yet, the update
  instructions can still be rewritten freely.
- FroniusSymoInverterActions moves to FroniusBatteryActions, bound to the
  FroniusBatteryHandler. Scripts have to retrieve the actions with the
  battery thing UID now.
- The firmware version lookup (/status/version, not device-scoped) moves to
  FroniusBaseThingHandler so that both the inverter handler (thing property)
  and the battery handler (battery control) can use it.
- The batterySettingsRefreshInterval parameter moves to the battery thing
  configuration (FroniusBatteryConfiguration).
- BATTERY_SETTINGS_CHANNELS moves to FroniusBindingConstants as per review.
