+++ bundles/org.openhab.binding.fronius/src/main/java/org/openhab/binding/fronius/internal/handler/FroniusBaseThingHandler.java
@@ -42,6 +42,8 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import com.google.gson.Gson;
+import com.google.gson.JsonElement;
+import com.google.gson.JsonParser;
 import com.google.gson.JsonSyntaxException;
 
 /**
@@ -186,6 +188,58 @@ public abstract class FroniusBaseThingHandler extends BaseThingHandler {
             throw new FroniusCommunicationException("Bridge handler is not available");
         }
         return bridgeHandler.getHttpUtil();
+    }
+
+    /**
+     * Get the firmware version of the inverter from its <code>/status/version</code> endpoint.
+     *
+     * @param scheme http or https
+     * @param hostname the hostname or IP address of the inverter
+     * @return the firmware version, or null if it could not be determined
+     */
+    protected @Nullable String getFirmwareVersion(String scheme, String hostname) {
+        final String host = scheme + "://" + hostname;
+        final String versionPath = "/status/version";
+        FroniusHttpUtil httpUtil;
+        try {
+            httpUtil = getHttpUtil();
+        } catch (FroniusCommunicationException e) {
+            logger.warn("Failed to get bridge HTTP utility for Fronius inverter at {}: {}", hostname, e.getMessage());
+            return null;
+        }
+
+        String url = host + "/api" + versionPath; // try the new API path first
+        String response;
+        try {
+            response = httpUtil.executeUrl(HttpMethod.GET, url, API_TIMEOUT);
+        } catch (FroniusCommunicationException e) {
+            url = host + versionPath; // fallback to the old API path
+            try {
+                response = httpUtil.executeUrl(HttpMethod.GET, url, API_TIMEOUT);
+            } catch (FroniusCommunicationException ex) {
+                logger.debug("Failed to get version info from Fronius inverter at {}: {}", hostname, ex.getMessage());
+                return null;
+            }
+        }
+        try {
+            JsonElement jsonElement = JsonParser.parseString(response);
+            if (!jsonElement.isJsonObject()) {
+                logger.debug("Invalid JSON response for version info from Fronius inverter at {}: {}", hostname,
+                        response);
+                return null;
+            }
+            try {
+                return jsonElement.getAsJsonObject().get("swrevisions").getAsJsonObject().get("GEN24").getAsString();
+            } catch (IllegalStateException | UnsupportedOperationException e) {
+                logger.debug("Failed to parse version info from Fronius inverter at {}: {}", hostname, e.getMessage());
+                return null;
+            }
+        } catch (JsonSyntaxException e) {
+            // 404 errors go here, as the response is not valid JSON
+            logger.debug("Invalid JSON response for version info from Fronius inverter at {}: {}", hostname, response,
+                    e);
+        }
+        return null;
     }
 
     /**
