+++ bundles/org.openhab.binding.fronius/src/main/java/org/openhab/binding/fronius/internal/handler/FroniusBatteryHandler.java
@@ -13,14 +13,24 @@
 package org.openhab.binding.fronius.internal.handler;
 
 import java.time.Instant;
+import java.util.Collection;
+import java.util.List;
+import java.util.Map;
 import java.util.Optional;
+import java.util.concurrent.ScheduledFuture;
+import java.util.concurrent.TimeUnit;
+import java.util.function.UnaryOperator;
 
 import org.eclipse.jdt.annotation.NonNullByDefault;
 import org.eclipse.jdt.annotation.Nullable;
-import org.openhab.binding.fronius.internal.FroniusBaseDeviceConfiguration;
+import org.openhab.binding.fronius.internal.FroniusBatteryConfiguration;
 import org.openhab.binding.fronius.internal.FroniusBindingConstants;
 import org.openhab.binding.fronius.internal.FroniusBridgeConfiguration;
+import org.openhab.binding.fronius.internal.action.FroniusBatteryActions;
+import org.openhab.binding.fronius.internal.api.FroniusBatteryControl;
+import org.openhab.binding.fronius.internal.api.FroniusBatteryControl.BatterySettings;
 import org.openhab.binding.fronius.internal.api.FroniusCommunicationException;
+import org.openhab.binding.fronius.internal.api.FroniusUnauthorizedException;
 import org.openhab.binding.fronius.internal.api.dto.storage.StorageController;
 import org.openhab.binding.fronius.internal.api.dto.storage.StorageDetails;
 import org.openhab.binding.fronius.internal.api.dto.storage.StorageRealtimeBody;
@@ -28,11 +38,18 @@ import org.openhab.binding.fronius.internal.api.dto.storage.StorageRealtimeBodyD
 import org.openhab.binding.fronius.internal.api.dto.storage.StorageRealtimeResponse;
 import org.openhab.core.library.types.DateTimeType;
 import org.openhab.core.library.types.DecimalType;
+import org.openhab.core.library.types.OnOffType;
 import org.openhab.core.library.types.QuantityType;
 import org.openhab.core.library.types.StringType;
 import org.openhab.core.library.unit.SIUnits;
 import org.openhab.core.library.unit.Units;
+import org.openhab.core.thing.Bridge;
+import org.openhab.core.thing.ChannelUID;
 import org.openhab.core.thing.Thing;
+import org.openhab.core.thing.binding.ThingHandlerService;
+import org.openhab.core.thing.firmware.types.SemverVersion;
+import org.openhab.core.types.Command;
+import org.openhab.core.types.RefreshType;
 import org.openhab.core.types.State;
 import org.openhab.core.types.UnDefType;
 import org.slf4j.Logger;
@@ -41,8 +58,12 @@ import org.slf4j.LoggerFactory;
 /**
  * The {@link FroniusBatteryHandler} polls the GetStorageRealtimeData endpoint and
  * maps fields from the Controller node to channels and thing properties.
+ * It also provides the battery settings channels and the battery control actions, which use the inverter's config API.
  *
  * @author Jimmy Tanagra - Initial contribution
+ * @author Christian Jonak-Möchel - Add battery settings channels and battery control actions
+ *
+ * @see FroniusBatteryActions
  */
 @NonNullByDefault
 public class FroniusBatteryHandler extends FroniusBaseThingHandler {
@@ -50,7 +71,11 @@ public class FroniusBatteryHandler extends FroniusBaseThingHandler {
     private final Logger logger = LoggerFactory.getLogger(FroniusBatteryHandler.class);
 
     private @Nullable StorageController controller;
-    private @Nullable FroniusBaseDeviceConfiguration config;
+    private @Nullable FroniusBatteryConfiguration config;
+    private @Nullable FroniusBatteryControl batteryControl;
+    private @Nullable BatterySettings lastBatterySettings;
+    private @Nullable Integer lastNightPreservationLimit;
+    private @Nullable ScheduledFuture<?> batterySettingsRefreshJob;
 
     public FroniusBatteryHandler(Thing thing) {
         super(thing);
@@ -63,24 +88,268 @@ public class FroniusBatteryHandler extends FroniusBaseThingHandler {
 
     @Override
     public void initialize() {
-        config = getConfigAs(FroniusBaseDeviceConfiguration.class);
+        FroniusBatteryConfiguration config = this.config = getConfigAs(FroniusBatteryConfiguration.class);
+        FroniusBridgeHandler bridgeHandler = getFroniusBridgeHandler();
+        Bridge bridge = getBridge();
+        if (bridge == null || bridgeHandler == null) {
+            logger.warn("bridge ({}) or bridgeHandler ({}) is null in initialize(), this is a bug, please report it.",
+                    bridge, bridgeHandler);
+            return;
+        }
+        FroniusBridgeConfiguration bridgeConfig = bridge.getConfiguration().as(FroniusBridgeConfiguration.class);
+        initializeBatteryControl(bridgeHandler, bridgeConfig.scheme, bridgeConfig.hostname, bridgeConfig.username,
+                bridgeConfig.password);
+        startBatterySettingsRefreshJob(config);
         super.initialize();
     }
 
     @Override
-    protected void handleRefresh(FroniusBridgeConfiguration bridgeConfiguration) throws FroniusCommunicationException {
-        FroniusBaseDeviceConfiguration config = this.config;
-        if (config == null) {
-            logger.warn("config is null in handleRefresh(), this is a bug, please report it.");
+    public Collection<Class<? extends ThingHandlerService>> getServices() {
+        return List.of(FroniusBatteryActions.class);
+    }
+
+    @Override
+    public void handleBridgeConfigurationUpdate(Map<String, Object> configurationParameters) {
+        super.handleBridgeConfigurationUpdate(configurationParameters);
+        Bridge bridge = getBridge();
+        FroniusBridgeHandler bridgeHandler = getFroniusBridgeHandler();
+        FroniusBatteryConfiguration config = this.config;
+        if (bridge == null || bridgeHandler == null) {
+            logger.warn(
+                    "bridge ({}) or bridgeHandler ({}) is null in handleBridgeConfigurationUpdate(), this is a bug, please report it.",
+                    bridge, bridgeHandler);
             return;
         }
-        updateData(bridgeConfiguration, config);
-        updateChannels();
-        updateProperties();
+        if (config == null) {
+            logger.warn("config is null in handleBridgeConfigurationUpdate(), this is a bug, please report it.");
+            return;
+        }
+        FroniusBridgeConfiguration bridgeConfig = bridge.getConfiguration().as(FroniusBridgeConfiguration.class);
+        initializeBatteryControl(bridgeHandler, bridgeConfig.scheme, bridgeConfig.hostname, bridgeConfig.username,
+                bridgeConfig.password);
+        startBatterySettingsRefreshJob(config);
+    }
+
+    @Override
+    public void dispose() {
+        cancelBatterySettingsRefreshJob();
+        super.dispose();
+    }
+
+    private void initializeBatteryControl(FroniusBridgeHandler bridgeHandler, String scheme, String hostname,
+            @Nullable String username, @Nullable String password) {
+        if (username == null || password == null) {
+            logger.info(
+                    "Credentials are not configured in the bridge. Battery control is not available for Thing '{}'.",
+                    thing.getUID());
+            return;
+        }
+
+        String firmwareVersion = getFirmwareVersion(scheme, hostname);
+        if (firmwareVersion == null) {
+            logger.warn(
+                    "The firmware version of the Fronius inverter could not be determined. Battery control is not available for Thing '{}'.",
+                    thing.getUID());
+            return;
+        }
+        int hyphenIndex = firmwareVersion.indexOf('-');
+        String versionString = (hyphenIndex > 0) ? firmwareVersion.substring(0, hyphenIndex) : firmwareVersion;
+        SemverVersion version = SemverVersion.fromString(versionString);
+        batteryControl = new FroniusBatteryControl(bridgeHandler.getConfigApiClient(), version, scheme, hostname,
+                username, password);
+    }
+
+    public @Nullable FroniusBatteryControl getBatteryControl() {
+        if (batteryControl == null) {
+            logger.warn("Battery control is not available for Thing '{}'. Check the bridge configuration.",
+                    thing.getUID());
+        }
+        return batteryControl;
+    }
+
+    @Override
+    public void handleCommand(ChannelUID channelUID, Command command) {
+        String channelId = channelUID.getIdWithoutGroup();
+        if (FroniusBindingConstants.BATTERY_SETTINGS_CHANNELS.contains(channelId)) {
+            FroniusBatteryControl control = getBatteryControl();
+            if (control == null) {
+                return;
+            }
+            try {
+                if (command instanceof RefreshType) {
+                    updateBatterySettingsChannels(control);
+                    refreshNightPreservationLimit(control);
+                    return;
+                }
+                switch (channelId) {
+                    case FroniusBindingConstants.BATTERY_CALIBRATION_CHANNEL,
+                            FroniusBindingConstants.BATTERY_NIGHT_PRESERVATION_LIMIT_CHANNEL -> {
+                        logger.warn("Channel {} is read-only, ignoring command {}", channelId, command);
+                        return;
+                    }
+                    case FroniusBindingConstants.BATTERY_CHARGE_FROM_GRID_CHANNEL -> {
+                        if (!(command instanceof OnOffType onOff)) {
+                            logger.warn("Unsupported command {} for channel {}", command, channelId);
+                            return;
+                        }
+                        boolean enabled = onOff == OnOffType.ON;
+                        control.setChargeFromGrid(enabled);
+                        applyUpdatedSettings(control, prev -> new BatterySettings(prev.minSoc(), prev.maxSoc(),
+                                prev.backupReservedCapacity(), prev.backupCriticalSoc(), enabled, prev.calibrating()));
+                    }
+                    default -> {
+                        int value;
+                        if (command instanceof QuantityType<?> quantity) {
+                            value = quantity.intValue();
+                        } else if (command instanceof DecimalType decimal) {
+                            value = decimal.intValue();
+                        } else {
+                            logger.warn("Unsupported command {} for channel {}", command, channelId);
+                            return;
+                        }
+                        switch (channelId) {
+                            case FroniusBindingConstants.BATTERY_BACKUP_RESERVED_CHANNEL -> {
+                                control.setBackupReservedCapacity(value);
+                                applyUpdatedSettings(control, prev -> new BatterySettings(prev.minSoc(), prev.maxSoc(),
+                                        value, prev.backupCriticalSoc(), prev.chargeFromGrid(), prev.calibrating()));
+                            }
+                            case FroniusBindingConstants.BATTERY_BACKUP_CRITICAL_SOC_CHANNEL -> {
+                                control.setBackupCriticalSoc(value);
+                                applyUpdatedSettings(control,
+                                        prev -> new BatterySettings(prev.minSoc(), prev.maxSoc(),
+                                                prev.backupReservedCapacity(), value, prev.chargeFromGrid(),
+                                                prev.calibrating()));
+                            }
+                            case FroniusBindingConstants.BATTERY_SOC_MIN_CHANNEL -> {
+                                BatterySettings settings = control.getBatterySettings();
+                                control.setSocLimits(value, settings.maxSoc());
+                                applyBatterySettings(new BatterySettings(value, settings.maxSoc(),
+                                        settings.backupReservedCapacity(), settings.backupCriticalSoc(),
+                                        settings.chargeFromGrid(), settings.calibrating()));
+                            }
+                            default -> {
+                                BatterySettings settings = control.getBatterySettings();
+                                control.setSocLimits(settings.minSoc(), value);
+                                applyBatterySettings(new BatterySettings(settings.minSoc(), value,
+                                        settings.backupReservedCapacity(), settings.backupCriticalSoc(),
+                                        settings.chargeFromGrid(), settings.calibrating()));
+                            }
+                        }
+                    }
+                }
+            } catch (FroniusCommunicationException | FroniusUnauthorizedException | IllegalArgumentException e) {
+                logger.warn("Failed to handle command for channel {}: {}", channelId, e.getMessage());
+            }
+            return;
+        }
+        super.handleCommand(channelUID, command);
+    }
+
+    private void updateBatterySettingsChannels(FroniusBatteryControl control)
+            throws FroniusCommunicationException, FroniusUnauthorizedException {
+        applyBatterySettings(control.getBatterySettings());
+    }
+
+    private void applyBatterySettings(BatterySettings settings) {
+        lastBatterySettings = settings;
+        updateState(FroniusBindingConstants.BATTERY_SOC_MIN_CHANNEL,
+                new QuantityType<>(settings.minSoc(), Units.PERCENT));
+        updateState(FroniusBindingConstants.BATTERY_SOC_MAX_CHANNEL,
+                new QuantityType<>(settings.maxSoc(), Units.PERCENT));
+        updateState(FroniusBindingConstants.BATTERY_BACKUP_RESERVED_CHANNEL,
+                new QuantityType<>(settings.backupReservedCapacity(), Units.PERCENT));
+        updateState(FroniusBindingConstants.BATTERY_BACKUP_CRITICAL_SOC_CHANNEL,
+                new QuantityType<>(settings.backupCriticalSoc(), Units.PERCENT));
+        updateState(FroniusBindingConstants.BATTERY_CHARGE_FROM_GRID_CHANNEL,
+                OnOffType.from(settings.chargeFromGrid()));
+        updateState(FroniusBindingConstants.BATTERY_CALIBRATION_CHANNEL, OnOffType.from(settings.calibrating()));
+    }
+
+    /**
+     * Updates the battery settings channels after a successful write: if the settings have been read before, the
+     * updated settings are derived from the known state to avoid a costly re-read from the config API; otherwise the
+     * settings are read from the inverter.
+     */
+    private void applyUpdatedSettings(FroniusBatteryControl control, UnaryOperator<BatterySettings> update)
+            throws FroniusCommunicationException, FroniusUnauthorizedException {
+        BatterySettings previous = lastBatterySettings;
+        if (previous != null) {
+            applyBatterySettings(update.apply(previous));
+        } else {
+            updateBatterySettingsChannels(control);
+        }
+    }
+
+    private void refreshNightPreservationLimit(FroniusBatteryControl control) {
+        if (!isLinked(FroniusBindingConstants.BATTERY_NIGHT_PRESERVATION_LIMIT_CHANNEL)) {
+            return;
+        }
+        try {
+            int limit = control.getNightPreservationLimit();
+            lastNightPreservationLimit = limit;
+            updateState(FroniusBindingConstants.BATTERY_NIGHT_PRESERVATION_LIMIT_CHANNEL,
+                    new QuantityType<>(limit, Units.PERCENT));
+        } catch (FroniusCommunicationException | FroniusUnauthorizedException e) {
+            logger.debug("Failed to read night preservation limit: {}", e.getMessage());
+        }
+    }
+
+    /**
+     * Periodically reads the battery settings to populate the battery settings channels, since they are not part of
+     * the cyclically polled Solar API data and reading them requires a costly login to the inverter's config API.
+     * The slow periodic refresh also picks up changes made through the inverter's web UI.
+     */
+    private void startBatterySettingsRefreshJob(FroniusBatteryConfiguration config) {
+        cancelBatterySettingsRefreshJob();
+        batterySettingsRefreshJob = scheduler.scheduleWithFixedDelay(() -> {
+            FroniusBatteryControl control = batteryControl;
+            if (control == null
+                    || FroniusBindingConstants.BATTERY_SETTINGS_CHANNELS.stream().noneMatch(this::isLinked)) {
+                return;
+            }
+            try {
+                updateBatterySettingsChannels(control);
+            } catch (FroniusCommunicationException | FroniusUnauthorizedException e) {
+                logger.warn("Failed to read battery settings: {}", e.getMessage());
+            }
+            refreshNightPreservationLimit(control);
+        }, 0, config.batterySettingsRefreshInterval, TimeUnit.MINUTES);
+    }
+
+    private void cancelBatterySettingsRefreshJob() {
+        ScheduledFuture<?> job = batterySettingsRefreshJob;
+        if (job != null) {
+            job.cancel(false);
+            batterySettingsRefreshJob = null;
+        }
     }
 
     @Override
     protected @Nullable State getValue(String channelId) {
+        BatterySettings settings = lastBatterySettings;
+        if (FroniusBindingConstants.BATTERY_SOC_MIN_CHANNEL.equals(channelId)) {
+            return settings == null ? null : new QuantityType<>(settings.minSoc(), Units.PERCENT);
+        }
+        if (FroniusBindingConstants.BATTERY_SOC_MAX_CHANNEL.equals(channelId)) {
+            return settings == null ? null : new QuantityType<>(settings.maxSoc(), Units.PERCENT);
+        }
+        if (FroniusBindingConstants.BATTERY_BACKUP_RESERVED_CHANNEL.equals(channelId)) {
+            return settings == null ? null : new QuantityType<>(settings.backupReservedCapacity(), Units.PERCENT);
+        }
+        if (FroniusBindingConstants.BATTERY_BACKUP_CRITICAL_SOC_CHANNEL.equals(channelId)) {
+            return settings == null ? null : new QuantityType<>(settings.backupCriticalSoc(), Units.PERCENT);
+        }
+        if (FroniusBindingConstants.BATTERY_CHARGE_FROM_GRID_CHANNEL.equals(channelId)) {
+            return settings == null ? null : OnOffType.from(settings.chargeFromGrid());
+        }
+        if (FroniusBindingConstants.BATTERY_CALIBRATION_CHANNEL.equals(channelId)) {
+            return settings == null ? null : OnOffType.from(settings.calibrating());
+        }
+        if (FroniusBindingConstants.BATTERY_NIGHT_PRESERVATION_LIMIT_CHANNEL.equals(channelId)) {
+            Integer limit = lastNightPreservationLimit;
+            return limit == null ? null : new QuantityType<>(limit, Units.PERCENT);
+        }
+
         StorageController local = controller;
         if (local == null) {
             return null;
@@ -118,7 +387,19 @@ public class FroniusBatteryHandler extends FroniusBaseThingHandler {
         };
     }
 
-    private void updateData(FroniusBridgeConfiguration bridgeConfiguration, FroniusBaseDeviceConfiguration config)
+    @Override
+    protected void handleRefresh(FroniusBridgeConfiguration bridgeConfiguration) throws FroniusCommunicationException {
+        FroniusBatteryConfiguration config = this.config;
+        if (config == null) {
+            logger.warn("config is null in handleRefresh(), this is a bug, please report it.");
+            return;
+        }
+        updateData(bridgeConfiguration, config);
+        updateChannels();
+        updateProperties();
+    }
+
+    private void updateData(FroniusBridgeConfiguration bridgeConfiguration, FroniusBatteryConfiguration config)
             throws FroniusCommunicationException {
         String location = FroniusBindingConstants.getBatteryDataUrl(bridgeConfiguration.scheme,
                 bridgeConfiguration.hostname, config.deviceId);
