+++ bundles/org.openhab.binding.fronius/src/main/java/org/openhab/binding/fronius/internal/handler/FroniusSymoInverterHandler.java
@@ -12,32 +12,16 @@
  */
 package org.openhab.binding.fronius.internal.handler;
 
-import static org.openhab.binding.fronius.internal.FroniusBindingConstants.API_TIMEOUT;
-
-import java.util.Collection;
-import java.util.List;
 import java.util.Map;
-import java.util.Set;
-import java.util.concurrent.ScheduledFuture;
-import java.util.concurrent.TimeUnit;
-import java.util.function.UnaryOperator;
 
 import javax.measure.Unit;
 
 import org.eclipse.jdt.annotation.NonNullByDefault;
 import org.eclipse.jdt.annotation.Nullable;
-import org.eclipse.jetty.client.HttpClient;
-import org.eclipse.jetty.http.HttpMethod;
 import org.openhab.binding.fronius.internal.FroniusBaseDeviceConfiguration;
 import org.openhab.binding.fronius.internal.FroniusBindingConstants;
 import org.openhab.binding.fronius.internal.FroniusBridgeConfiguration;
-import org.openhab.binding.fronius.internal.FroniusSymoInverterConfiguration;
-import org.openhab.binding.fronius.internal.action.FroniusSymoInverterActions;
-import org.openhab.binding.fronius.internal.api.FroniusBatteryControl;
-import org.openhab.binding.fronius.internal.api.FroniusBatteryControl.BatterySettings;
 import org.openhab.binding.fronius.internal.api.FroniusCommunicationException;
-import org.openhab.binding.fronius.internal.api.FroniusHttpUtil;
-import org.openhab.binding.fronius.internal.api.FroniusUnauthorizedException;
 import org.openhab.binding.fronius.internal.api.dto.ValueUnit;
 import org.openhab.binding.fronius.internal.api.dto.inverter.*;
 import org.openhab.binding.fronius.internal.api.dto.powerflow.PowerFlowRealtimeBody;
@@ -50,19 +34,10 @@ import org.openhab.core.library.types.OnOffType;
 import org.openhab.core.library.types.QuantityType;
 import org.openhab.core.library.unit.Units;
 import org.openhab.core.thing.Bridge;
-import org.openhab.core.thing.ChannelUID;
 import org.openhab.core.thing.Thing;
-import org.openhab.core.thing.binding.ThingHandlerService;
-import org.openhab.core.thing.firmware.types.SemverVersion;
-import org.openhab.core.types.Command;
-import org.openhab.core.types.RefreshType;
 import org.openhab.core.types.State;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-
-import com.google.gson.JsonElement;
-import com.google.gson.JsonParser;
-import com.google.gson.JsonSyntaxException;
 
 /**
  * The {@link FroniusSymoInverterHandler} is responsible for updating the data, which are
@@ -77,28 +52,15 @@ import com.google.gson.JsonSyntaxException;
 @NonNullByDefault
 public class FroniusSymoInverterHandler extends FroniusBaseThingHandler {
 
-    private static final Set<String> BATTERY_SETTINGS_CHANNELS = Set.of(FroniusBindingConstants.BATTERY_SOC_MIN_CHANNEL,
-            FroniusBindingConstants.BATTERY_SOC_MAX_CHANNEL, FroniusBindingConstants.BATTERY_BACKUP_RESERVED_CHANNEL,
-            FroniusBindingConstants.BATTERY_BACKUP_CRITICAL_SOC_CHANNEL,
-            FroniusBindingConstants.BATTERY_CHARGE_FROM_GRID_CHANNEL,
-            FroniusBindingConstants.BATTERY_CALIBRATION_CHANNEL,
-            FroniusBindingConstants.BATTERY_NIGHT_PRESERVATION_LIMIT_CHANNEL);
-
     private final Logger logger = LoggerFactory.getLogger(FroniusSymoInverterHandler.class);
-    private final HttpClient httpClient;
 
     private @Nullable InverterRealtimeResponse inverterRealtimeResponse;
     private @Nullable PowerFlowRealtimeResponse powerFlowResponse;
-    private @Nullable FroniusSymoInverterConfiguration config;
+    private @Nullable FroniusBaseDeviceConfiguration config;
     private @Nullable InverterInfo inverterInfo;
-    private @Nullable FroniusBatteryControl batteryControl;
-    private @Nullable BatterySettings lastBatterySettings;
-    private @Nullable Integer lastNightPreservationLimit;
-    private @Nullable ScheduledFuture<?> batterySettingsRefreshJob;
 
-    public FroniusSymoInverterHandler(Thing thing, HttpClient httpClient) {
+    public FroniusSymoInverterHandler(Thing thing) {
         super(thing);
-        this.httpClient = httpClient;
     }
 
     @Override
@@ -108,39 +70,13 @@ public class FroniusSymoInverterHandler extends FroniusBaseThingHandler {
 
     @Override
     protected void handleRefresh(FroniusBridgeConfiguration bridgeConfiguration) throws FroniusCommunicationException {
-        FroniusSymoInverterConfiguration config = this.config;
+        FroniusBaseDeviceConfiguration config = this.config;
         if (config == null) {
             logger.warn("config is null in handleRefresh(), this is a bug, please report it.");
             return;
         }
         updateData(bridgeConfiguration, config);
         updateChannels();
-    }
-
-    private void initializeBatteryControl(FroniusBridgeHandler bridgeHandler, String scheme, String hostname,
-            @Nullable String username, @Nullable String password) {
-        if (username == null || password == null) {
-            logger.info(
-                    "Credentials are not configured in the bridge. Battery control is not available for Thing '{}'.",
-                    thing.getUID());
-            return;
-        }
-
-        InverterInfo localInverterInfo = inverterInfo;
-        if (localInverterInfo != null) {
-            String firmwareVersion = localInverterInfo.firmware();
-            if (firmwareVersion != null) {
-                int hyphenIndex = firmwareVersion.indexOf('-');
-                String versionString = (hyphenIndex > 0) ? firmwareVersion.substring(0, hyphenIndex) : firmwareVersion;
-                SemverVersion version = SemverVersion.fromString(versionString);
-                batteryControl = new FroniusBatteryControl(bridgeHandler.getConfigApiClient(), version, scheme,
-                        hostname, username, password);
-                return;
-            }
-        }
-        logger.warn(
-                "The firmware version of the Fronius inverter could not be determined. Battery control is not available for Thing '{}'.",
-                thing.getUID());
     }
 
     private void updateProperties() {
@@ -160,38 +96,25 @@ public class FroniusSymoInverterHandler extends FroniusBaseThingHandler {
 
     @Override
     public void initialize() {
-        FroniusSymoInverterConfiguration config = this.config = getConfigAs(FroniusSymoInverterConfiguration.class);
-        FroniusBridgeHandler bridgeHandler = getFroniusBridgeHandler();
+        FroniusBaseDeviceConfiguration config = this.config = getConfigAs(FroniusBaseDeviceConfiguration.class);
         Bridge bridge = getBridge();
-        if (bridge == null || bridgeHandler == null) {
-            logger.warn("bridge ({}) or bridgeHandler ({}) is null in initialize(), this is a bug, please report it.",
-                    bridge, bridgeHandler);
+        if (bridge == null) {
+            logger.warn("bridge is null in initialize(), this is a bug, please report it.");
             return;
         }
         FroniusBridgeConfiguration bridgeConfig = bridge.getConfiguration().as(FroniusBridgeConfiguration.class);
         inverterInfo = getInverterInfo(bridgeConfig.scheme, bridgeConfig.hostname, config.deviceId);
         updateProperties();
-        initializeBatteryControl(bridgeHandler, bridgeConfig.scheme, bridgeConfig.hostname, bridgeConfig.username,
-                bridgeConfig.password);
-        startBatterySettingsRefreshJob(config);
         super.initialize();
-    }
-
-    @Override
-    public Collection<Class<? extends ThingHandlerService>> getServices() {
-        return List.of(FroniusSymoInverterActions.class);
     }
 
     @Override
     public void handleBridgeConfigurationUpdate(Map<String, Object> configurationParameters) {
         super.handleBridgeConfigurationUpdate(configurationParameters);
         Bridge bridge = getBridge();
-        FroniusBridgeHandler bridgeHandler = getFroniusBridgeHandler();
-        FroniusSymoInverterConfiguration config = this.config;
-        if (bridge == null || bridgeHandler == null) {
-            logger.warn(
-                    "bridge ({}) or bridgeHandler ({}) is null in handleBridgeConfigurationUpdate(), this is a bug, please report it.",
-                    bridge, bridgeHandler);
+        FroniusBaseDeviceConfiguration config = this.config;
+        if (bridge == null) {
+            logger.warn("bridge is null in handleBridgeConfigurationUpdate(), this is a bug, please report it.");
             return;
         }
         if (config == null) {
@@ -201,179 +124,6 @@ public class FroniusSymoInverterHandler extends FroniusBaseThingHandler {
         FroniusBridgeConfiguration bridgeConfig = bridge.getConfiguration().as(FroniusBridgeConfiguration.class);
         inverterInfo = getInverterInfo(bridgeConfig.scheme, bridgeConfig.hostname, config.deviceId);
         updateProperties();
-        initializeBatteryControl(bridgeHandler, bridgeConfig.scheme, bridgeConfig.hostname, bridgeConfig.username,
-                bridgeConfig.password);
-        startBatterySettingsRefreshJob(config);
-    }
-
-    public @Nullable FroniusBatteryControl getBatteryControl() {
-        if (batteryControl == null) {
-            logger.warn("Battery control is not available for Thing '{}'. Check the bridge configuration.",
-                    thing.getUID());
-        }
-        return batteryControl;
-    }
-
-    @Override
-    public void handleCommand(ChannelUID channelUID, Command command) {
-        String channelId = channelUID.getIdWithoutGroup();
-        if (BATTERY_SETTINGS_CHANNELS.contains(channelId)) {
-            FroniusBatteryControl control = getBatteryControl();
-            if (control == null) {
-                return;
-            }
-            try {
-                if (command instanceof RefreshType) {
-                    updateBatterySettingsChannels(control);
-                    refreshNightPreservationLimit(control);
-                    return;
-                }
-                switch (channelId) {
-                    case FroniusBindingConstants.BATTERY_CALIBRATION_CHANNEL,
-                            FroniusBindingConstants.BATTERY_NIGHT_PRESERVATION_LIMIT_CHANNEL -> {
-                        logger.warn("Channel {} is read-only, ignoring command {}", channelId, command);
-                        return;
-                    }
-                    case FroniusBindingConstants.BATTERY_CHARGE_FROM_GRID_CHANNEL -> {
-                        if (!(command instanceof OnOffType onOff)) {
-                            logger.warn("Unsupported command {} for channel {}", command, channelId);
-                            return;
-                        }
-                        boolean enabled = onOff == OnOffType.ON;
-                        control.setChargeFromGrid(enabled);
-                        applyUpdatedSettings(control, prev -> new BatterySettings(prev.minSoc(), prev.maxSoc(),
-                                prev.backupReservedCapacity(), prev.backupCriticalSoc(), enabled, prev.calibrating()));
-                    }
-                    default -> {
-                        int value;
-                        if (command instanceof QuantityType<?> quantity) {
-                            value = quantity.intValue();
-                        } else if (command instanceof DecimalType decimal) {
-                            value = decimal.intValue();
-                        } else {
-                            logger.warn("Unsupported command {} for channel {}", command, channelId);
-                            return;
-                        }
-                        switch (channelId) {
-                            case FroniusBindingConstants.BATTERY_BACKUP_RESERVED_CHANNEL -> {
-                                control.setBackupReservedCapacity(value);
-                                applyUpdatedSettings(control, prev -> new BatterySettings(prev.minSoc(), prev.maxSoc(),
-                                        value, prev.backupCriticalSoc(), prev.chargeFromGrid(), prev.calibrating()));
-                            }
-                            case FroniusBindingConstants.BATTERY_BACKUP_CRITICAL_SOC_CHANNEL -> {
-                                control.setBackupCriticalSoc(value);
-                                applyUpdatedSettings(control,
-                                        prev -> new BatterySettings(prev.minSoc(), prev.maxSoc(),
-                                                prev.backupReservedCapacity(), value, prev.chargeFromGrid(),
-                                                prev.calibrating()));
-                            }
-                            case FroniusBindingConstants.BATTERY_SOC_MIN_CHANNEL -> {
-                                BatterySettings settings = control.getBatterySettings();
-                                control.setSocLimits(value, settings.maxSoc());
-                                applyBatterySettings(new BatterySettings(value, settings.maxSoc(),
-                                        settings.backupReservedCapacity(), settings.backupCriticalSoc(),
-                                        settings.chargeFromGrid(), settings.calibrating()));
-                            }
-                            default -> {
-                                BatterySettings settings = control.getBatterySettings();
-                                control.setSocLimits(settings.minSoc(), value);
-                                applyBatterySettings(new BatterySettings(settings.minSoc(), value,
-                                        settings.backupReservedCapacity(), settings.backupCriticalSoc(),
-                                        settings.chargeFromGrid(), settings.calibrating()));
-                            }
-                        }
-                    }
-                }
-            } catch (FroniusCommunicationException | FroniusUnauthorizedException | IllegalArgumentException e) {
-                logger.warn("Failed to handle command for channel {}: {}", channelId, e.getMessage());
-            }
-            return;
-        }
-        super.handleCommand(channelUID, command);
-    }
-
-    private void updateBatterySettingsChannels(FroniusBatteryControl control)
-            throws FroniusCommunicationException, FroniusUnauthorizedException {
-        applyBatterySettings(control.getBatterySettings());
-    }
-
-    private void applyBatterySettings(BatterySettings settings) {
-        lastBatterySettings = settings;
-        updateState(FroniusBindingConstants.BATTERY_SOC_MIN_CHANNEL,
-                new QuantityType<>(settings.minSoc(), Units.PERCENT));
-        updateState(FroniusBindingConstants.BATTERY_SOC_MAX_CHANNEL,
-                new QuantityType<>(settings.maxSoc(), Units.PERCENT));
-        updateState(FroniusBindingConstants.BATTERY_BACKUP_RESERVED_CHANNEL,
-                new QuantityType<>(settings.backupReservedCapacity(), Units.PERCENT));
-        updateState(FroniusBindingConstants.BATTERY_BACKUP_CRITICAL_SOC_CHANNEL,
-                new QuantityType<>(settings.backupCriticalSoc(), Units.PERCENT));
-        updateState(FroniusBindingConstants.BATTERY_CHARGE_FROM_GRID_CHANNEL,
-                OnOffType.from(settings.chargeFromGrid()));
-        updateState(FroniusBindingConstants.BATTERY_CALIBRATION_CHANNEL, OnOffType.from(settings.calibrating()));
-    }
-
-    /**
-     * Updates the battery settings channels after a successful write: if the settings have been read before, the
-     * updated settings are derived from the known state to avoid a costly re-read from the config API; otherwise the
-     * settings are read from the inverter.
-     */
-    private void applyUpdatedSettings(FroniusBatteryControl control, UnaryOperator<BatterySettings> update)
-            throws FroniusCommunicationException, FroniusUnauthorizedException {
-        BatterySettings previous = lastBatterySettings;
-        if (previous != null) {
-            applyBatterySettings(update.apply(previous));
-        } else {
-            updateBatterySettingsChannels(control);
-        }
-    }
-
-    private void refreshNightPreservationLimit(FroniusBatteryControl control) {
-        if (!isLinked(FroniusBindingConstants.BATTERY_NIGHT_PRESERVATION_LIMIT_CHANNEL)) {
-            return;
-        }
-        try {
-            int limit = control.getNightPreservationLimit();
-            lastNightPreservationLimit = limit;
-            updateState(FroniusBindingConstants.BATTERY_NIGHT_PRESERVATION_LIMIT_CHANNEL,
-                    new QuantityType<>(limit, Units.PERCENT));
-        } catch (FroniusCommunicationException | FroniusUnauthorizedException e) {
-            logger.debug("Failed to read night preservation limit: {}", e.getMessage());
-        }
-    }
-
-    /**
-     * Periodically reads the battery settings to populate the battery settings channels, since they are not part of
-     * the cyclically polled Solar API data and reading them requires a costly login to the inverter's config API.
-     * The slow periodic refresh also picks up changes made through the inverter's web UI.
-     */
-    private void startBatterySettingsRefreshJob(FroniusSymoInverterConfiguration config) {
-        cancelBatterySettingsRefreshJob();
-        batterySettingsRefreshJob = scheduler.scheduleWithFixedDelay(() -> {
-            FroniusBatteryControl control = batteryControl;
-            if (control == null || BATTERY_SETTINGS_CHANNELS.stream().noneMatch(this::isLinked)) {
-                return;
-            }
-            try {
-                updateBatterySettingsChannels(control);
-            } catch (FroniusCommunicationException | FroniusUnauthorizedException e) {
-                logger.warn("Failed to read battery settings: {}", e.getMessage());
-            }
-            refreshNightPreservationLimit(control);
-        }, 0, config.batterySettingsRefreshInterval, TimeUnit.MINUTES);
-    }
-
-    private void cancelBatterySettingsRefreshJob() {
-        ScheduledFuture<?> job = batterySettingsRefreshJob;
-        if (job != null) {
-            job.cancel(false);
-            batterySettingsRefreshJob = null;
-        }
-    }
-
-    @Override
-    public void dispose() {
-        cancelBatterySettingsRefreshJob();
-        super.dispose();
     }
 
     /**
@@ -384,30 +134,6 @@ public class FroniusSymoInverterHandler extends FroniusBaseThingHandler {
      */
     @Override
     protected @Nullable State getValue(String channelId) {
-        BatterySettings settings = lastBatterySettings;
-        if (FroniusBindingConstants.BATTERY_SOC_MIN_CHANNEL.equals(channelId)) {
-            return settings == null ? null : new QuantityType<>(settings.minSoc(), Units.PERCENT);
-        }
-        if (FroniusBindingConstants.BATTERY_SOC_MAX_CHANNEL.equals(channelId)) {
-            return settings == null ? null : new QuantityType<>(settings.maxSoc(), Units.PERCENT);
-        }
-        if (FroniusBindingConstants.BATTERY_BACKUP_RESERVED_CHANNEL.equals(channelId)) {
-            return settings == null ? null : new QuantityType<>(settings.backupReservedCapacity(), Units.PERCENT);
-        }
-        if (FroniusBindingConstants.BATTERY_BACKUP_CRITICAL_SOC_CHANNEL.equals(channelId)) {
-            return settings == null ? null : new QuantityType<>(settings.backupCriticalSoc(), Units.PERCENT);
-        }
-        if (FroniusBindingConstants.BATTERY_CHARGE_FROM_GRID_CHANNEL.equals(channelId)) {
-            return settings == null ? null : OnOffType.from(settings.chargeFromGrid());
-        }
-        if (FroniusBindingConstants.BATTERY_CALIBRATION_CHANNEL.equals(channelId)) {
-            return settings == null ? null : OnOffType.from(settings.calibrating());
-        }
-        if (FroniusBindingConstants.BATTERY_NIGHT_PRESERVATION_LIMIT_CHANNEL.equals(channelId)) {
-            Integer limit = lastNightPreservationLimit;
-            return limit == null ? null : new QuantityType<>(limit, Units.PERCENT);
-        }
-
         final String[] fields = channelId.split("#");
         if (fields.length < 1) {
             return null;
@@ -419,7 +145,7 @@ public class FroniusSymoInverterHandler extends FroniusBaseThingHandler {
             return null;
         }
 
-        FroniusSymoInverterConfiguration config = this.config;
+        FroniusBaseDeviceConfiguration config = this.config;
         if (config == null) {
             logger.warn("config is null in getValue(String channelId), this is a bug, please report it.");
             return null;
@@ -684,49 +410,7 @@ public class FroniusSymoInverterHandler extends FroniusBaseThingHandler {
         }
 
         final String serial = String.valueOf(data.getUniqueID());
-
-        final String host = scheme + "://" + ip;
-        final String versionPath = "/status/version";
-        FroniusHttpUtil httpUtil;
-        try {
-            httpUtil = getHttpUtil();
-        } catch (FroniusCommunicationException e) {
-            logger.warn("Failed to get bridge HTTP utility for Fronius inverter at {}: {}", ip, e.getMessage());
-            return null;
-        }
-
-        String url = host + "/api" + versionPath; // try the new API path first
-        String response;
-        try {
-            response = httpUtil.executeUrl(HttpMethod.GET, url, API_TIMEOUT);
-        } catch (FroniusCommunicationException e) {
-            url = host + versionPath; // fallback to the old API path
-            try {
-                response = httpUtil.executeUrl(HttpMethod.GET, url, API_TIMEOUT);
-            } catch (FroniusCommunicationException ex) {
-                logger.debug("Failed to get version info from Fronius inverter at {}: {}", ip, ex.getMessage());
-                return new InverterInfo(serial, null);
-            }
-        }
-        try {
-            JsonElement jsonElement = JsonParser.parseString(response);
-            if (!jsonElement.isJsonObject()) {
-                logger.debug("Invalid JSON response for version info from Fronius inverter at {}: {}", ip, response);
-                return new InverterInfo(serial, null);
-            }
-            try {
-                String firmware = jsonElement.getAsJsonObject().get("swrevisions").getAsJsonObject().get("GEN24")
-                        .getAsString();
-                return new InverterInfo(serial, firmware);
-            } catch (IllegalStateException | UnsupportedOperationException e) {
-                logger.debug("Failed to parse version info from Fronius inverter at {}: {}", ip, e.getMessage());
-                return new InverterInfo(serial, null);
-            }
-        } catch (JsonSyntaxException e) {
-            // 404 errors go here, as the response is not valid JSON
-            logger.debug("Invalid JSON response for version info from Fronius inverter at {}: {}", ip, response, e);
-        }
-        return new InverterInfo(serial, null);
+        return new InverterInfo(serial, getFirmwareVersion(scheme, ip));
     }
 
     private record InverterInfo(String serial, @Nullable String firmware) {
